test refactor: remove test with broken assumption
To remove a user from the admin list doesn't matter, even though its
clearly something one may expect should matter.

The issue is that JupyterHub itself makes use of the admin_users list as
a way to declare new admins, not to remove users having been marked as
admin in the past.

What to do about this is not clear, but having a test for this doesn't
make sense.